requirements.txt
streamlit>=1.36.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
pandas>=2.0.0
